Refactor dialogs into separate files
